Move (old) checkpoint creation logic to `Cardano.Wallet.Checkpoints`
